% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\docType{package}
\name{fit}
\alias{fit}
\alias{fit-package}
\title{fit: pre-built model for the gestational diabetes project}
\description{
fit: pre-built model for the gestational diabetes project
}
\section{fit object}{

The fit object is a xgbTree class
}

