% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_models.R
\name{load_models}
\alias{load_models}
\title{Load models pre-built for the gestational diabetes project}
\usage{
load_models()
}
\value{
a list objects named models.list which has 6 elements
}
\description{
Load models pre-built for the gestational diabetes project
}
\examples{
load_models()
names(models.list)
}
